requirements.txt
discord
gpt-2-simple
asyncio
python-dotenv
flask